Refactor and fix label/annotation remover
The changes have become more generically a shim for the metadata in api
responses
