+++ autocrud/apply.go
@@ -71,14 +71,10 @@ func serverSideApply(ctx context.Context, clientGetter KubernetesClientGetter, a
 	responseManifest := res.UnstructuredContent()
 	id := createID(responseManifest)
 
-	// remove internal labels and annotations not set in config
 	responseMetadata := responseManifest["metadata"].(map[string]any)
 	configMetadata := manifest["metadata"].(map[string]any)
-	removeInternalKeys(responseMetadata["labels"].(map[string]any), configMetadata["labels"].(map[string]any))
-	removeInternalKeys(responseMetadata["annotations"].(map[string]any), configMetadata["annotations"].(map[string]any))
-	// remove regex matching labels and annotations specified by the user in the provider block
-	removeKeys(responseMetadata["labels"].(map[string]any), configMetadata["labels"].(map[string]any), clientGetter.IgnoreLabels())
-	removeKeys(responseMetadata["annotations"].(map[string]any), configMetadata["annotations"].(map[string]any), clientGetter.IgnoreAnnotations())
+
+	shimMetadata(responseMetadata, configMetadata, clientGetter.IgnoreLabels(), clientGetter.IgnoreAnnotations())
 
 	err = FlattenManifest(responseManifest, model)
 	if err != nil {
